Cloud-Automation/GCP/LAB04-Cloud-Functions-Deployment/requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
google-cloud-storage==2.13.0
pytest==7.4.3 